C4

+++ main.c
@@ -3,8 +3,8 @@
 int main()
 {
     puts("Введите пожалуйста два числа, а я выведу вам сумму");
-    int a, b;
-    scanf("%d %d", &a, &b);
-    printf("%d\n", a + b);
+    float a, b;
+    scanf("%f %f", &a, &b);
+    printf("%f\n", a + b);
     return 0;
 }
